Update the test cases all
